+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TODO:
 
 pre-0.29.2:
 - add `stepperConnectToPin(pin, driver_type)` to allow the module selection
+- fix issue #170 and use consequently `fas_max()` instead of `max()`
 
 0.29.1:
 - improve rounding for log2/pow2 conversion